Add deploy config for Consul
So far it happened multiple times that Consul container was killed with error code 137 (Container using too much memory).

This commit attempts to fix the problem by configuring automatic container restart on such failures.

In addition, this also adds strict reservation and limit for memory and cpu to see the effect when running consul instance.

(The goal is to set resources as minimum as possible so that it can run on cheapest VMs)