Add reverse proxy - nginx setup
This commit adds nginx reverse proxy setup that forward requests to dedicated server allowing them to stay unreachable directly.

Nginx will serve as the gateway that runs on public instance, receives all public traffic and forwarding to dedicated servers.

The checked in configuration is cloned from default one:

```
$ docker run --rm --entrypoint=cat nginx /etc/nginx/nginx.conf > components/nginx/nginx.conf
```

and is desired to be adjusted to the project needs.